fix(benchmark): fix construct of synthese bench tests for admin (#3548)
The function `add_bluring_to_benchmark_test_class` - used in "test_benchmark_synthese.py" - does not only to add tests for user_with_blurring, but it also finalizes the declaration of the synthese benchmark tests for the user admin_user. But, before the current commit, the `"user_profile"` was unwishingly changed from `admin_user` to `user_with_blurring` before the declaration of the tests for admin_user, thus Breaking the execution of those lasts tests.